backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.28.0
github-copilot-sdk>=0.1.23
sse-starlette>=2.2.0
pydantic>=2.0
